# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Blog api documentation.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Fetch all tags.",
                "operationId": "49b8aae23d92f665318c3273cb54ec75",
                "responses": {
                    "200": {
                        "description": "Fetch all tags"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tags"
                ],
                "summary": "Create a new tag",
                "operationId": "9e21dd3b04a3e4b23f12a3ce0de8c7e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag created successfully."
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tags/{id}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Fetch the tag with specific id.",
                "operationId": "ef5174291db360e7b51137941ba41425",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the tag to be fetched.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag fetched successfully"
                    },
                    "404": {
                        "description": "Tag not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tags"
                ],
                "summary": "Update tag with specified id.",
                "operationId": "e7eae54e99e2e29945c84eae30df1350",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the tag to be updated.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Tag updated successfully"
                    },
                    "404": {
                        "description": "Tag not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tags"
                ],
                "summary": "Delete tag with specified id.",
                "operationId": "b3c3d74fb2ee76e77f08e3fc9dc80a10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the tag to be deleted.",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tag deleted successfully."
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tags"
                ],
                "summary": "Update tag with specified id.",
                "operationId": "009b5a15b3518a154b7ba1eaabd5189e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id for the tag to be updated.",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tag"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Tag updated successfully"
                    },
                    "404": {
                        "description": "Tag not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tag": {
                "title": "Create a Tag",
                "properties": {
                    "title": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}